+++ exporters/console.py
@@ -7,7 +7,10 @@ from core.event import MCPEvent
 from utils.logger import logger
 
 class ConsoleExporter(BaseExporter):
+
     def export(self, event: MCPEvent) -> None:
         status_icon = "✅" if event.status == "success" else "❌"
         logger.info(f"{status_icon} [Exec time: {event.delta}] {event.tool_name} - Status: {event.status}, Args: {event.args}, Result: {event.result}")
-        
+    
+    def export_batch(self, event_batch):
+        return super().export_batch(event_batch)